requirements.txt
mcp>=1.0.0,
httpx>=0.25.0,
pydantic>=2.0.0,
python-dotenv>=1.0.0,
uvicorn>=0.25.0,
requests>=2.31.0,
fastmcp>=2.14.1